feat: add protected route group for authenticated layout
