+++ apps/learner/src/routes/(protected)/(core)/+layout.svelte
