+++ apps/learner/src/routes/(protected)/(core)/explore/+page.svelte
