+++ apps/learner/src/routes/(protected)/collection/[id]/+page.svelte
