+++ apps/learner/src/routes/(protected)/content/[id]/+page.server.ts
