+++ apps/learner/src/routes/(protected)/content/[id]/+page.svelte
