+++ apps/learner/src/routes/(protected)/content/[id]/quiz/+page.server.ts
