+++ apps/learner/src/routes/(protected)/content/[id]/quiz/+page.svelte
